requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0
pdf2image>=1.16.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Local LLM support
ctransformers>=0.2.0
llama-cpp-python>=0.2.0
accelerate>=0.20.0

# Data processing and validation
pydantic>=2.0.0
jsonschema>=4.17.0
pyyaml>=6.0

# Web framework for API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0
